MDL-39958 core_group: Bulk unassign of groupings uses low-level API
As the previous commit, we keep the legacy event for now as it would
be wrong to create a new event to replace it.

Also, the code has been changed to call the low-level API to unassign
groups from groupins, even though at the moment there are no
events for that function.