MDL-40419 backup: Restore non-multiple blocks with different settings
Some blocks are not allowed to be duplicated in a certain context,
however when we restored a block that is meant to be displayed on
sub pages, but already exists on the main context, it was not restored.
Now we would first check if the existing block shares the same config
and if not we will duplicate the block.