MDL-57543 ajax: Bulk Ajax requests when possible
If a few requests are sent within a short amount of time, we will
bulk them into a single request to improve performance and reduce
load on the server. This is transparent to the user and developer.

If all scheduled requests were not requiring login, then the
entry point will be service-nologin.php. However, if only one
of the scheduled requests requires login, then all of them will
be directed to service.php.